Add CI script for running Clippy
CI script will run clippy with `RUSTFLAGS="--deny warnings"` because
this is how we do it in the other build CI scripts. However this might
clash with configs that are done in the `.cargo/config` file.
I expect this will not be a significant problem however and the CI
script is not a critical resource and as such it seems like a waste to
bikeshed this.